feat: Enable JSX transpilation with Babel
Integrates Babel standalone into index.html to allow transpilation of JSX code directly in the browser. This simplifies local development by removing the need for a separate build step for JSX.

The script type in index.html has been updated to `text/babel` with `data-type="module"` to facilitate this change.